Refactor ProfileControllerIT into separate modules and fix tests
